+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/CourseRestController.java
@@ -34,6 +34,9 @@ public class CourseRestController {
         return courseController.getCourseById(courseId);
     }
 
-
+    @DeleteMapping("/{courseId}")
+    public void deleteCourse(@PathVariable Long courseId){
+        courseController.deleteCourse(courseId);
+    }
 
 }