+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/StudentRestController.java
@@ -31,6 +31,12 @@ public class StudentRestController {
         return studentController.getStudentByCourseId(courseId);
     }
 
+    @GetMapping("")
+    public List<Student> getAllStudents() {
+        return studentController.getAllStudents();
+    }
+
+
 
 
 }