+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/StudentController.java
@@ -38,4 +38,8 @@ public class StudentController {
         Course course = courseService.getCourseById(courseId);
         return course.getStudents();
     }
+
+    public List<Student> getAllStudents() {
+        return studentService.getAllStudents();
+    }
 }