+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/CourseService.java
@@ -28,4 +28,8 @@ public class CourseService {
     public Course saveCourse(Course course) {
         return courseRepository.save(course);
     }
+
+    public void deleteCourse(Course course) {
+        courseRepository.delete(course);
+    }
 }