+++ NIMIO2024_TFG_PolMorenoCastelar/changelog.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - [NPP-21] 》CO13F - Add a student to a course with only id's
 - [NPP-23] 》CO15F - Get all students and course details by courseID
 - [NPP-24] 》CO16F - Get only students of a course by courseID
+- [NPP-25] 》CO17F - Delete a course by courseID without deleting students
 
 
 - [NPP-44] 》ST38F - Create a student associated with a course